Removed multiple queues for SPI

+++ ros/src/spi_bridge/core/spi_brige_common.h
@@ -14,5 +14,5 @@ constexpr int LOOP_RATE = 30; //processing frequency
 
 
 // HW setup
-constexpr int SPI_QUEUE_SIZE = 8192; // Size of inbound/outbound queue
-constexpr int SPI_INT_QUEUE_SIZE = 6000; // Size of interrupt queue
+constexpr size_t SPI_QUEUE_SIZE = 8192; // Size of inbound/outbound queue
+constexpr size_t SPI_INT_QUEUE_SIZE = 6000; // Size of interrupt queue
